refactor: move AI spacing modules from utils/ into ai-spacing/
content-script.ts and utils/ were getting harder to read as the AI
spacing feature grew inside a generic utils/ folder. Moves the four
AI spacing modules (ai-spacing.ts, hyphen-prompt.ts, hyphen-sign.ts,
prompt-classifier.ts) into their own ai-spacing/ folder, renaming the
contract file to ambiguous-shape.ts since it defines both halves of
an ambiguous shape (AmbiguousShape and PromptSpec). Updates imports,
tests, and docs path mentions accordingly. No behavior change.
